refactor(editor): move record-form to components
