+++ libs/feature/editor/src/index.ts
@@ -5,7 +5,7 @@ export * from './lib/+state/editor.reducer'
 export * from './lib/+state/editor.actions'
 export * from './lib/feature-editor.module'
 export * from './lib/services/editor.service'
-export * from './lib/record-form/record-form.component'
+export * from './lib/components/record-form/record-form.component'
 export * from './lib/components/wizard/wizard.component'
 export * from './lib/components/wizard-field/wizard-field.component'
 export * from './lib/components/wizard-summarize/wizard-summarize.component'
